requirements.txt
interactions.py>=4.4.0
Pillow>=8.0.0
numpy>=1.19.0
opencv-python>=4.5.0
python-dotenv>=1.0.0
pytest>=7.0.0
black>=24.0.0
isort>=5.0.0
psutil>=5.9.0